fix: вернул обратно классовый компонент, пофиксил баг с ЛК
