+++ src/components/Article/indexbackup.js
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { useEffect, Suspence } from 'react';
 import { connect } from 'react-redux';
-import marked from 'marked';
 import ArticleMeta from './ArticleMeta';
 import CommentContainer from './CommentContainer';
 import agent from '../../agent';
 import { ARTICLE_PAGE_LOADED, ARTICLE_PAGE_UNLOADED } from '../../constants/actionTypes';
+import Post from './Post';
 
 const mapStateToProps = (state) => ({
   ...state.article,
@@ -12,84 +12,91 @@ const mapStateToProps = (state) => ({
 });
 
 const mapDispatchToProps = (dispatch) => ({
-  onLoad: (payload) => dispatch({ type: ARTICLE_PAGE_LOADED, payload }),
+  onLoad: (payload) => console.log(dispatch({ type: ARTICLE_PAGE_LOADED, payload })),
   onUnload: () => dispatch({ type: ARTICLE_PAGE_UNLOADED }),
 });
 
-class Article extends React.Component {
-  componentWillMount() {
-    this.props.onLoad(Promise.all([
-      agent.Articles.get(this.props.match.params.id),
-      agent.Comments.forArticle(this.props.match.params.id),
+const Article = (props) => {
+  const {
+    onLoad, onUnload, match, article = { kek: 'anus' }, currentUser = {}, comments = [], commentErrors,
+  } = props;
+
+  const {
+    body,
+    author = {},
+    title,
+    tagList = [],
+  } = article;
+
+  const { id: matchId } = match.params;
+  const currentUserUsername = currentUser?.username;
+  const authorUsername = author?.username;
+  // const { Articles, Comments } = agent;
+
+  useEffect(() => {
+    onLoad(Promise.all([
+      agent.Articles.get(props.match.params.id),
+      agent.Comments.forArticle(props.match.params.id),
     ]));
-  }
 
-  componentWillUnmount() {
-    this.props.onUnload();
-  }
+    return () => onUnload();
+  }, []);
 
-  render() {
-    if (!this.props.article) {
-      return null;
-    }
+  const canModify = currentUser
+    && currentUserUsername === authorUsername;
 
-    const markup = { __html: marked(this.props.article.body, { sanitize: true }) };
-    const canModify = this.props.currentUser
-      && this.props.currentUser.username === this.props.article.author.username;
-    return (
-      <div className="article-page">
+  return article && (
+    <div className="article-page">
 
-        <div className="banner">
-          <div className="container">
+      <div className="banner">
+        <div className="container">
 
-            <h1>{this.props.article.title}</h1>
-            <ArticleMeta
-              article={this.props.article}
-              canModify={canModify}
-            />
+          <ArticleMeta
+            article={article}
+            canModify={canModify}
+          />
 
-          </div>
-        </div>
-
-        <div className="container page">
-
-          <div className="row article-content">
-            <div className="col-xs-12">
-
-              <div dangerouslySetInnerHTML={markup} />
-
-              <ul className="tag-list">
-                {
-      this.props.article.tagList.map((tag) => (
-        <li
-          className="tag-default tag-pill tag-outline"
-          key={tag}
-        >
-          {tag}
-        </li>
-      ))
-    }
-              </ul>
-
-            </div>
-          </div>
-
-          <hr />
-
-          <div className="article-actions" />
-
-          <div className="row">
-            <CommentContainer
-              comments={this.props.comments || []}
-              errors={this.props.commentErrors}
-              slug={this.props.match.params.id}
-              currentUser={this.props.currentUser}
-            />
-          </div>
         </div>
       </div>
-    );
-  }
-}
+
+      <div className="container page">
+
+        <div className="row article-content">
+          <div className="col-xs-12">
+
+            <Post body={body} title={title} />
+
+            <ul className="tag-list">
+              {
+                tagList?.map((tag) => (
+                  <li
+                    className="tag-default tag-pill tag-outline"
+                    key={tag}
+                  >
+                    {tag}
+                  </li>
+                ))
+              }
+            </ul>
+
+          </div>
+        </div>
+
+        <hr />
+
+        <div className="article-actions" />
+
+        <div className="row">
+          <CommentContainer
+            comments={comments}
+            errors={commentErrors}
+            slug={matchId}
+            currentUser={currentUser}
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
 
 export default connect(mapStateToProps, mapDispatchToProps)(Article);
