+++ src/components/Header.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-
 import icons from './UI/icons/icons';
 
 const {
